Allow plugins to override the vanilla commands
While plugins could already override vanilla commands, it wasn't 100% perfect. (Example: Making a plugin register a "ban" command will have buggy arguments because it is still using the old "vanilla" command arguments. This commit checks if there is a command conflict with any of the vanilla commands, and, if there is, it is automagically removed from the knownCommands map.